backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
python-multipart
pdfplumber
python-docx
python-jose[cryptography] 
passlib[bcrypt] 
fastapi[all]
faiss-cpu
sentence-transformers
langchain
langchain_community
langchain_core
langchain_groq
langchain_openai
langchain-huggingface

